# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANAM-BF - API de la plateforme agrométéo",
    "description": "\nAPI de la plateforme agrométéorologique de l'ANAM-BF.\n\n**Conventions**\n- Format JSON, UTF-8. Dates en UTC au format ISO 8601.\n- Authentification : en-tête `Authorization: Bearer <access_token>` (voir la section *Authentification*).\n  Les contenus publiés (bulletins, alertes, avis, carte, zones) sont **publics** : consultables sans compte.\n- Erreurs : `{\"detail\": \"message lisible\"}` avec le code HTTP adapté (401 non connecté, 403 droits insuffisants,\n  404 introuvable, 409 conflit d'état, 422 données invalides, 429 trop de tentatives).\n- Listes paginées : paramètres `limit` et `offset`, réponse `{\"total\": N, \"items\": [...]}`.\n- Langues : `fr` (français), `en` (anglais), `mos` (mooré).\n- Génération audio/vidéo : asynchrone. Lancer, puis relire le contenu jusqu'à `media.status` = `ready`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "List Users",
        "description": "**Accès :** Connexion requise, permission `users:read` (rôles : Agent ANAM, Administrateur).",
        "operationId": "list_users_admin_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ex. pending = comptes à valider",
              "title": "Status"
            },
            "description": "ex. pending = comptes à valider"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "commune",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Commune"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "recherche dans nom, e-mail, téléphone, identifiant",
              "title": "Q"
            },
            "description": "recherche dans nom, e-mail, téléphone, identifiant"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Create Internal Account",
        "description": "Crée un compte (agent ANAM, administrateur, responsable communal…), actif immédiatement.\nL'utilisateur devra changer son mot de passe initial à sa première connexion.\n\n**Accès :** Connexion requise, permission `users:manage` (rôles : Administrateur).",
        "operationId": "create_internal_account_admin_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalAccountCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Get User",
        "description": "**Accès :** Connexion requise, permission `users:read` (rôles : Agent ANAM, Administrateur).",
        "operationId": "get_user_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/role": {
      "patch": {
        "tags": [
          "Administration"
        ],
        "summary": "Change Role",
        "description": "**Accès :** Connexion requise, permission `roles:assign` (rôles : Administrateur).",
        "operationId": "change_role_admin_users__user_id__role_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/activate": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Activate User",
        "description": "Valide un compte en attente (ex. compte local) ou réactive un compte désactivé.\n\n**Accès :** Connexion requise, permission `users:manage` (rôles : Administrateur).",
        "operationId": "activate_user_admin_users__user_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/disable": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Disable User",
        "description": "Désactive un compte et ferme toutes ses sessions.\n\n**Accès :** Connexion requise, permission `users:manage` (rôles : Administrateur).",
        "operationId": "disable_user_admin_users__user_id__disable_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/reset-2fa": {
      "post": {
        "tags": [
          "Administration"
        ],
        "summary": "Reset Two Factor",
        "description": "Réinitialise la 2FA (téléphone perdu) : l'utilisateur devra la reconfigurer à sa prochaine connexion.\n\n**Accès :** Connexion requise, permission `users:manage` (rôles : Administrateur).",
        "operationId": "reset_two_factor_admin_users__user_id__reset_2fa_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/stats/communes": {
      "get": {
        "tags": [
          "Administration"
        ],
        "summary": "Users By Commune",
        "description": "Répartition des comptes actifs par commune.\n\n**Accès :** Connexion requise, permission `stats:read` (rôles : Responsable communal, Agent ANAM, Administrateur).",
        "operationId": "users_by_commune_admin_stats_communes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommuneStat"
                  },
                  "type": "array",
                  "title": "Response Users By Commune Admin Stats Communes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Channel": {
        "type": "string",
        "enum": [
          "sms",
          "push",
          "email"
        ],
        "title": "Channel"
      },
...
      "CommuneStat": {
        "properties": {
          "commune": {
            "type": "string",
            "title": "Commune"
          },
          "users": {
            "type": "integer",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "commune",
          "users"
        ],
        "title": "CommuneStat"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InternalAccountCreate": {
        "properties": {
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "commune": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune"
          }
        },
        "type": "object",
        "required": [
          "password",
          "role"
        ],
        "title": "InternalAccountCreate",
        "description": "Création d'un compte agent / administrateur / etc. par un administrateur."
      },
      "Language": {
        "type": "string",
        "enum": [
          "fr",
          "mos",
          "en"
        ],
        "title": "Language"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "grand_public",
          "responsable_communal",
          "agent_anam",
          "observateur",
          "administrateur"
        ],
        "title": "Role"
      },
      "RoleChange": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "RoleChange"
      },
...
      "UserList": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "UserList"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "commune": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune"
          },
          "language": {
            "$ref": "#/components/schemas/Language"
          },
          "notification_channels": {
            "items": {
              "$ref": "#/components/schemas/Channel"
            },
            "type": "array",
            "title": "Notification Channels"
          },
          "totp_enabled": {
            "type": "boolean",
            "title": "Totp Enabled"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone",
          "email",
          "username",
          "full_name",
          "role",
          "status",
          "commune",
          "language",
          "notification_channels",
          "totp_enabled",
          "must_change_password",
          "created_at",
          "last_login_at"
        ],
        "title": "UserOut"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "disabled"
        ],
        "title": "UserStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}